requirements.txt
anyio==4.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
darkdetect==0.7.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
metabase-api==3.5.1
numpy==2.3.1
oauthlib==3.3.1
pandas==2.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyqtdarktheme==2.1.0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
shiboken6==6.9.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
